Move away from alpha blending
With the addition of ambient light, alpha blending isn't making as much
sense, as "no ambient light" was still darker, due to the blending of
alpha values.

Now we're back to operating on the main texture directly, which gives us
a lot more control with ambient light.

Now when we set a brightness to 0.0, the scene is black, and when we set
it to 100.0, the brightness is at full, which is a much more sane API in
my books.