Remove requirement to have ambient light
Lights not showing up unless a camera has an ambient light component
isn't the best feeling, especially from a perspective of "let's try out
this new library".

While ambient light is likely going to be used most of the time, there's
no strict requirement for it to be present to spawn lights.