Add HDR support
As HDR uses a different texture format, we need to set it on a per view
basis.

As the pipeline is currently more of a "one and done" construct, we need to
make use of a specialized pipeline. It allows us to dynamically create
the render pipeline descriptor, giving us a place to hook into if each
view has HDR enabled.

On the view node side of things, we need to ensure we're grabbing the
specialized pipeline ID from the view entity, as we are no longer
constructing a "static" variant of the pipeline.