Write "empty" point light to render world
GpuArrayBuffer won't write to the GPU if we have no point lights (an
empty array).

We first tried writing a more custom integration with the GPU, but ended
up running into a very similar problem.

For now lets use GPUArrayBuffer and all it does for us, writing an "empty"
point light to ensure we'll always have a binding.