Add light occlusion (#20)
* Adds LightOccluder2d component

* Adds SDF Pass

* Use fixed size uniform buffer in WebGL2

Dynamic arrays aren't supported in WebGL2, so we have to fall back to
a fixed size array in a uniform binding.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Almost setup ray marching? :)

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Kill bevy_pbr

PBR is for 3D stuff. It's nice to not need it.

We'll want to clean up our view conversion definitions in the future.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Add isolated light map pass

The resulting texture isn't used yet, but it's being written to.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Use light map

Now we're using our light map, blended with our main texture.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Fix ambient light

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Check occluder visibility

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Use separate node for SDF passo

Not strictly necessary, but I do like having an explicit render graph,
especially if it's something someone else ever wanted to tweak.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Use enum for light occluder shapes

This will allow us to sanely expand our API as we introduce more
occluder shapes.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Pull out light map node

A refactor, again getting us to a world of a proper render graph.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Move sdf shader to sdf module

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Use SDF specific component for texture

Nice to split this one up, rather than all the textures living on one
component.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Add separate light map texture component

Now we can completely isolate our light map and sdf textures.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Hoist types shader

This shader isn't specific to any one pipeline, so let's pull it up.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Use shared shader types

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Remove lighting settings

No blur yet. Will add this back later.

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Fix comment

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Pull view transformations to separate shader

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Add occluders to changelog

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Add light occlusion example

Co-Authored-By: Miguel Albernaz <albernazmiguel@gmail.com>

* Fix occluder field order

These were wrong!

* Add more occluders to occlusion example

* Use texture sample level

Missed this when porting things over.

* Fix ambient light

* Split up shape from light occluder

This will allow us to encode more information into an occluder, such as
information on how shadows should be cast.

* Allow shadows to be toggled on a per light basis

* Update docs

* Remove future goals

---------

Co-authored-by: Miguel Albernaz <albernazmiguel@gmail.com>