Bind point light counts
Once space is allocated in a buffer, it's allocated. Despawning a light
doesn't clear the memory in the binding, unless we make a new binding.

Prior to this change, despawning a light could have it hang around, still
in memory.

This change adds a separate binding to track the number of active point
lights, and uses the number to iterate through the point light buffer,
ensuring we won't over step.