Add an "empty" buffer
When our GpuArrayBuffer has no data in it, we don't have a binding to use
to send to the GPU, in which case the node generally won't run.

We can use this "empty" buffer to supply our shaders with a "junk" binding,
allowing the shaders to still run (assuming we don't read from this).