Use extracted occluder count to build SDF texture
The way GPUArrayBuffer works, there may be allocated space that does not
hold useful data.

This is fine if you never despawn an occluder, but as soon as we do, we
end up with "junk" data that might still exist in the buffer. If we use
arrayLength to determine the number of occluders in play, we're not
going to get the right answer.

Instead, we can bind the number of occluders separately, and use that
for an accurate count.