Require marker component for processing lights
Originally I wanted to make this library as easy to use as possible. This
meant that all `Camera2d` components would process lighting out of the
box.

However, this has lead to a few instances of "the principle of least
surprise" being broken.

For example, a UI only camera is something that currently isn't possible
with the existing library.

An option I considered was an opt-out component, but I don't super feel
this aligns with "the Bevy way". I view an opt-in as something less
likely to cause problems / surprises down the road.