Fix occluders not despawning correctly (#47)
* Extract occluder count

Right now we're using the allocated GPU space to determine the number
of occluders, which is wrong.

To fix that, first let's extra the number of occluders.

* Use extracted occluder count to build SDF texture

The way GPUArrayBuffer works, there may be allocated space that does not
hold useful data.

This is fine if you never despawn an occluder, but as soon as we do, we
end up with "junk" data that might still exist in the buffer. If we use
arrayLength to determine the number of occluders in play, we're not
going to get the right answer.

Instead, we can bind the number of occluders separately, and use that
for an accurate count.

* Use max texture value when no occluders exist

If there aren't any occluders, we can default to the max texture value
(which maps to distance).

This change fixes #43, as we're no longer attempting to act on junk
data.

* Update CHANGELOG.md